+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Use compact help template (which allows to show help for just one command).
+* Fix wrong export target when just one resource is specified (#11).
 
 ## 0.5.1 (2018-07-12)